requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
emoji==2.2.0
et-xmlfile==1.1.0
flake8==5.0.4
greenlet==2.0.2
idna==3.4
importlib-metadata==4.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
numpy==1.21.6
openpyxl==3.1.2
pandas==1.3.5
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0